FIX  : Mypage & Ranking 연결 작업

+++ fin-point/src/main/java/com/fp/finpoint/web/mypage/MypageDto.java
@@ -20,7 +20,7 @@ public class MypageDto {
 
     public Long spend;
 
-    public Long pieceKind;
+    public int pieceKind;
 
     public Long pieceCnt;
